refactor: consolidate error handling logic across files
- Add a new file `errorHandler.js` for a general error handler
- Import `errorHandler.js` in `order.api.test.js` to use the error handler
- Remove the duplicate error handling code from `order.api.test.js`
