+++ tests/k6/order.api.test.js
@@ -13,39 +13,7 @@
 
 import http from 'k6/http';
 import {check, group, sleep} from 'k6';
-
-// General error handler to log error details.
-class ErrorHandler {
-  // Instruct the error handler how to log errors
-  constructor(logErrorDetails) {
-    this.logErrorDetails = logErrorDetails;
-  }
-
-  // Logs response error details if isError is true.
-  logError(isError, res, tags = {}) {
-    if (!isError) return;
-
-    // the Traceparent header is a W3C Trace Context
-    const traceparentHeader = res.request.headers['Traceparent'];
-
-    // Add any other useful information
-    const errorData = Object.assign(
-        {
-          url: res.url,
-          status: res.status,
-          error_code: res.error_code,
-          traceparent: traceparentHeader && traceparentHeader.toString(),
-        },
-        tags,
-    );
-    this.logErrorDetails(errorData);
-  }
-}
-
-// Set up the error handler to log errors to the console
-const errorHandler = new ErrorHandler((error) => {
-  console.error(error);
-});
+import errorHandler from './errorHandler.js';
 
 const scenarios = {
   average_load: {
